[FEATURE] data.py Module for parse Python dictionary in special and load into database as apps or users. The format of the dictionary is described as pydoc format in the py files.
[FEATURE] ffos.scripts package. A package that have scripts to
work independently from django but are important to the
application. It may use django recources.

[FEATURE] ffos.scripts.loaddata. A script to load data to the
database. It work for apps or users. The script documentation is
in itself as pydoc.

[BUGFIX] ffos.models. Some firefox data had some fields at null
that the data base wouldn't allow null. This is a problem because
the data files structure lack documentation. It's fix for now, at
least for this data files.

[MINOR] ffos.settings A middleware for transactions were added.